requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
celery
redis
requests